pyproject.toml
[project]
name = "parr-prompt"
version = "0.1.0"
description = "Send prompts to OpenAI, Gemini, Grok, Claude, DeepSeek, Kimi, or Doubao (Volcengine Ark)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.40.0",
    "anthropic>=0.34.0",
    "google-genai>=1.0.0",
    "dash>=2.14.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
parr-prompt = "parr_prompt.cli:main"
parr-web = "parr_prompt.web_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["parr_prompt"]
